+++ package.json
@@ -1,82 +1,82 @@
 {
 	"name": "tldraw-server",
 	"type": "module",
-  "version": "0.0.1",
-  "license": "AGPL-3.0",
-  "description": "tldraw backend",
-  "author": "Schulcloud-Verbund",
-  "private": true,
+	"version": "0.0.1",
+	"license": "AGPL-3.0",
+	"description": "tldraw backend",
+	"author": "Schulcloud-Verbund",
+	"private": true,
 	"engines": {
-    "npm": ">=8.0.0",
+		"npm": ">=8.0.0",
 		"node": ">=20.0.0"
 	},
-  "scripts": {
-    "build": "nest build",
-    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "start": "nest start server",
-    "start:server:dev": "nest start server --watch",
-    "start:server:debug": "nest start server --debug --watch",
-    "start:server:prod": "node  --env-file .env dist/apps/tldraw-server.js",
-    "start:worker:dev": "nest start worker --watch",
-    "start:worker:debug": "nest start worker --debug --watch",
-    "start:worker:prod": "node  --env-file .env dist/apps/tldraw-worker.js",
-    "lint": "eslint \"{src}/**/*.ts\" --fix",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:cov": "jest --coverage",
-    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand"
-  },
-  "dependencies": {
-    "@nestjs/common": "^10.4.1",
-    "@nestjs/config": "^3.2.3",
-    "@nestjs/core": "^10.4.1",
-    "@nestjs/platform-express": "^10.4.1",
-    "@y/redis": "github:hpi-schul-cloud/y-redis#deb646c062f0d743111831abc4cd406dad48f880",
-    "ioredis": "^5.4.1",
-    "prom-client": "^15.1.3",
-    "reflect-metadata": "^0.2.2",
-    "rxjs": "^7.8.1",
-    "uws": "github:uNetworking/uWebSockets.js#v20.47.0",
-    "y-protocols": "^1.0.6"
-  },
-  "devDependencies": {
-    "@nestjs/cli": "^10.4.4",
-    "@nestjs/schematics": "^10.1.3",
-    "@nestjs/testing": "^10.4.1",
-    "@types/express": "^4.17.21",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.3.1",
-    "@types/supertest": "^6.0.0",
-    "@typescript-eslint/eslint-plugin": "^7.0.0",
-    "@typescript-eslint/parser": "^7.0.0",
-    "eslint": "^8.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "jest": "^29.7.0",
-    "prettier": "^3.3.3",
-    "source-map-support": "^0.5.21",
-    "supertest": "^7.0.0",
-    "ts-jest": "^29.2.4",
-    "ts-loader": "^9.5.1",
-    "ts-node": "^10.9.2",
-    "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.5.4"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "ts"
-    ],
-    "rootDir": "src",
-    "testRegex": ".*\\.spec\\.ts$",
-    "transform": {
-      "^.+\\.(t|j)s$": "ts-jest"
-    },
-    "collectCoverageFrom": [
-      "**/*.(t|j)s"
-    ],
-    "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
-  }
+	"scripts": {
+		"build": "nest build",
+		"format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+		"start": "nest start server",
+		"start:server:dev": "nest start server --watch",
+		"start:server:debug": "nest start server --debug --watch",
+		"start:server:prod": "node  --env-file .env dist/apps/tldraw-server.js",
+		"start:worker:dev": "nest start worker --watch",
+		"start:worker:debug": "nest start worker --debug --watch",
+		"start:worker:prod": "node  --env-file .env dist/apps/tldraw-worker.js",
+		"lint": "eslint \"{src}/**/*.ts\" --fix",
+		"test": "jest",
+		"test:watch": "jest --watch",
+		"test:cov": "jest --coverage",
+		"test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand"
+	},
+	"dependencies": {
+		"@nestjs/common": "^10.4.1",
+		"@nestjs/config": "^3.2.3",
+		"@nestjs/core": "^10.4.1",
+		"@nestjs/platform-express": "^10.4.1",
+		"@y/redis": "github:hpi-schul-cloud/y-redis#c9863e74d8d22dd75f53e894835cede66afdddd9",
+		"ioredis": "^5.4.1",
+		"prom-client": "^15.1.3",
+		"reflect-metadata": "^0.2.2",
+		"rxjs": "^7.8.1",
+		"uws": "github:uNetworking/uWebSockets.js#v20.47.0",
+		"y-protocols": "^1.0.6"
+	},
+	"devDependencies": {
+		"@nestjs/cli": "^10.4.4",
+		"@nestjs/schematics": "^10.1.3",
+		"@nestjs/testing": "^10.4.1",
+		"@types/express": "^4.17.21",
+		"@types/jest": "^29.5.12",
+		"@types/node": "^20.3.1",
+		"@types/supertest": "^6.0.0",
+		"@typescript-eslint/eslint-plugin": "^7.0.0",
+		"@typescript-eslint/parser": "^7.0.0",
+		"eslint": "^8.0.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-prettier": "^5.2.1",
+		"jest": "^29.7.0",
+		"prettier": "^3.3.3",
+		"source-map-support": "^0.5.21",
+		"supertest": "^7.0.0",
+		"ts-jest": "^29.2.4",
+		"ts-loader": "^9.5.1",
+		"ts-node": "^10.9.2",
+		"tsconfig-paths": "^4.2.0",
+		"typescript": "^5.5.4"
+	},
+	"jest": {
+		"moduleFileExtensions": [
+			"js",
+			"json",
+			"ts"
+		],
+		"rootDir": "src",
+		"testRegex": ".*\\.spec\\.ts$",
+		"transform": {
+			"^.+\\.(t|j)s$": "ts-jest"
+		},
+		"collectCoverageFrom": [
+			"**/*.(t|j)s"
+		],
+		"coverageDirectory": "../coverage",
+		"testEnvironment": "node"
+	}
 }
