+++ src/infra/authorization/authorization.module.mts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { AuthorizationService } from './authorization.service.mjs';
+
+@Module({
+    providers: [AuthorizationService],
+    exports: [AuthorizationService],
+})
+export class AuthorizationModule {}
