+++ src/infra/authorization/index.mts
@@ -1,0 +1,3 @@
+export * from './authorization.module.mjs';
+export * from './authorization.service.mjs';
+
