+++ src/infra/redis/index.mts
@@ -1,0 +1,3 @@
+export * from './redis.module.mjs';
+export * from './redis.service.mjs';
+
