+++ src/infra/redis/redis.module.mts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { RedisService } from './redis.service.mjs';
+
+@Module({
+    providers: [RedisService],
+    exports: [RedisService],
+})
+export class RedisModule {}
