+++ src/infra/storage/index.mts
@@ -1,0 +1,3 @@
+export * from './storage.module.mjs';
+export * from './storage.service.mjs';
+
