+++ src/infra/storage/storage.module.mts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { StorageService } from './storage.service.mjs';
+
+@Module({
+    providers: [StorageService],
+    exports: [StorageService],
+})
+export class StorageModule {}
